Fix network req/resp tests
