+++ packages/lodestar/src/network/nodejs/bundle.ts
@@ -1,5 +1,5 @@
 /**
- * @module network/libp2p/nodejs
+ * @module network/nodejs
  */
 
 import LibP2p from "libp2p";
