+++ packages/lodestar/test/unit/network/util.ts
@@ -1,5 +1,5 @@
-import {NodejsNode} from "../../../../src/network/libp2p/nodejs";
-import {createPeerId, initializePeerInfo} from "../../../../src/network/libp2p/util";
+import {NodejsNode} from "../../../src/network/nodejs";
+import {createPeerId, initializePeerInfo} from "../../../src/network/util";
 
 export async function createNode(multiaddr: string): Promise<NodejsNode> {
   const peerId = await createPeerId();
